+++ app/src/main/java/development/iarratais/fragment/MemoryFragment.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2016 Iarratais Development
+ *
+ * karl.development@gmail.com
+ */
+
+package development.iarratais.fragment;
+
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import development.iarratais.androidinfo.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class MemoryFragment extends Fragment {
+
+    View rootView;
+
+    public MemoryFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        rootView = inflater.inflate(R.layout.fragment_memory, container, false);
+
+        return rootView;
+    }
+
+}