+++ changelog.md
@@ -1,4 +1,11 @@
-## AndroidInfo Changelog
+## My Device Information Changelog
+
+#2
++ Added APKMirror link in the play services tab
++ Added Ads Lite library
+* Enabled minify
++ Check if mobile data is available on the device (targeted at tablets)
++ Ads in the play services and network tab
 
 #1.20160724
 + Added main interface